+++ keywords.txt
@@ -9,11 +9,11 @@
 #######################################
 
 CosmosNV2	KEYWORD1
-SensorDeLinha	KEYWORD1
+SensorDeRefletancia	KEYWORD1
 SensorDeToque	KEYWORD1
 SensorDeDistancia	KEYWORD1
 SensorDeCor	KEYWORD1
-Buzzer	KEYWORD1
+AltoFalante	KEYWORD1
 Motor	KEYWORD1
 ServoMotor	KEYWORD1
 Tela	KEYWORD1
@@ -26,8 +26,8 @@ ler	KEYWORD2
 lerEstado	KEYWORD2
 iniciar	KEYWORD2
 obterValores	KEYWORD2
-soar_bipe	KEYWORD2
-soar_dois_bipes	KEYWORD2
+bipar	KEYWORD2
+tocar	KEYWORD2
 configurarVelocidade	KEYWORD2
 girar	KEYWORD2
 parar	KEYWORD2
@@ -58,6 +58,8 @@ A9	LITERAL1	Constants
 A10	LITERAL1	Constants
 B	LITERAL1	Constants
 C	LITERAL1	Constants
+D	LITERAL1	Constants
+E	LITERAL1	Constants
 L1	LITERAL1	Constants
 L2	LITERAL1	Constants
 L3	LITERAL1	Constants
@@ -65,8 +67,6 @@ L4	LITERAL1	Constants
 L5	LITERAL1	Constants
 L6	LITERAL1	Constants
 L7	LITERAL1	Constants
-D	LITERAL1	Constants
-E	LITERAL1	Constants
 M1	LITERAL1	Constants
 M2	LITERAL1	Constants
 S1	LITERAL1	Constants
