+++ src/Constantes.h
@@ -1,0 +1,29 @@
+
+#define B 1
+#define C 2
+#define D 3
+#define E 4
+
+#define M1 1
+#define M2 2
+
+#define PARA_FRENTE  1
+#define PARA_TRAS    2
+
+#define L1 1
+#define L2 2
+#define L3 3
+#define L4 4
+#define L5 5
+#define L6 6
+#define L7 7
+
+#define S1 1
+#define S2 2
+#define S3 3
+#define S4 4
+#define S5 5
+
+#define PRETO   0
+#define BRANCO  1
+#define INVERSO 2
