+++ src/Pinagem.h
@@ -1,0 +1,40 @@
+
+#define PINO_SDA_SENSOR_DE_COR_DIREITO    42
+#define PINO_SCL_SENSOR_DE_COR_DIREITO    44
+#define PINO_SDA_SENSOR_DE_COR_ESQUERDO   38
+#define PINO_SCL_SENSOR_DE_COR_ESQUERDO   40
+
+#define PINO_PORTA_B 18
+#define PINO_PORTA_C 36
+
+#define PINO_PORTA_L1 23
+#define PINO_PORTA_L2 25
+#define PINO_PORTA_L3 27
+#define PINO_PORTA_L4 29
+#define PINO_PORTA_L5 31
+#define PINO_PORTA_L6 33
+#define PINO_PORTA_L7 35
+
+#define PINO_XSHUT_SENSOR_PORTA_L1 PINO_PORTA_L1
+#define PINO_XSHUT_SENSOR_PORTA_L2 PINO_PORTA_L2
+#define PINO_XSHUT_SENSOR_PORTA_L3 PINO_PORTA_L3
+#define PINO_XSHUT_SENSOR_PORTA_L4 PINO_PORTA_L4
+#define PINO_XSHUT_SENSOR_PORTA_L5 PINO_PORTA_L5
+#define PINO_XSHUT_SENSOR_PORTA_L6 PINO_PORTA_L6
+#define PINO_XSHUT_SENSOR_PORTA_L7 PINO_PORTA_L7
+
+#define PINO_VELOCIDADE_MOTOR_1   7
+#define PINO_MOTOR1_A             6
+#define PINO_MOTOR1_B             5
+
+#define PINO_MOTOR2_A             4
+#define PINO_MOTOR2_B             3
+#define PINO_VELOCIDADE_MOTOR_2   2
+
+#define PINO_PORTA_S1 12
+#define PINO_PORTA_S2 11
+#define PINO_PORTA_S3 10
+#define PINO_PORTA_S4 9
+#define PINO_PORTA_S5 8
+
+#define PINO_ALTO_FALANTE 17
