+++ src/SensorDeCor.cpp
@@ -1,60 +1,8 @@
-/* ------------------------------------------------------------------------------------------ *\
-|                                                                                              |
-|                                                                                              |
-|                                    Cosmos NV2 Library                                        |
-|                                                                                              |
-|                                                                                              |
-\* ------------------------------------------------------------------------------------------ */
-
-/*
-
-Author: Vinícios Silva
-Meu GitHub: github.com/ViniciosSilva811
-Versão: 2.0 Fevereiro de 2020
-
-*/
-
-// ----------------------------------- DEPENDÊNCIAS ----------------------------------------- //
 
 #include <Arduino.h>
-#include "CosmosNV2.h"
-#include "deps/SoftwareWire/SoftwareWire.h"
-
-// ------------------------------------ CONSTANTES ------------------------------------------ //
-
-#define ENDERECO_L1 0x31
-#define ENDERECO_L2 0x32
-#define ENDERECO_L3 0x33
-#define ENDERECO_L4 0x34
-#define ENDERECO_L5 0x35
-#define ENDERECO_L6 0x36
-#define ENDERECO_L7 0x37
-
-// #define SHUT_L1 14
-// #define SHUT_L2 15
-// #define SHUT_L3 16
-// #define SHUT_L4 17
-// #define SHUT_L5 19
-// #define SHUT_L6 24
-// #define SHUT_L7 22
-
-#define SHUT_L1 23
-#define SHUT_L2 25
-#define SHUT_L3 27
-#define SHUT_L4 29
-#define SHUT_L5 31
-#define SHUT_L6 33
-#define SHUT_L7 35
-
-// #define SDAD 48                                     // Pino SDA do sensor de cor direito
-// #define SCLD 46                                     // Pino SCL do sensor de cor direito
-// #define SDAE 50                                     // Pino SDA do sensor do cor esquerdo
-// #define SCLE 52                                     // Pino SCL do sensor de cor esquerdo
-
-#define SDAE 38                                     // Pino SDA do sensor do cor esquerdo
-#define SCLE 40                                     // Pino SCL do sensor de cor esquerdo
-#define SDAD 42                                     // Pino SDA do sensor de cor direito
-#define SCLD 44                                     // Pino SCL do sensor de cor direito
+#include "SensorDeCor.h"
+#include "Constantes.h"
+#include "Pinagem.h"
 
 #define TCS34725_ADDRESS          (0x29)
 
@@ -107,109 +55,18 @@ Versão: 2.0 Fevereiro de 2020
 #define TCS34725_BDATAL           (0x1A)    /* Blue channel data */
 // #define TCS34725_BDATAH           (0x1B)
 
-#define BUZZER 17 // A0
-
-#define PINO_VELOCIDADE_MOTOR_1 7
-#define MOTOR1_A 6
-#define MOTOR1_B 5
-
-#define MOTOR2_A 4
-#define MOTOR2_B 3
-#define PINO_VELOCIDADE_MOTOR_2 2
-
-// ------------------------------- Classe SensorDeLinha ------------------------------------- //
-
-SensorDeLinha::SensorDeLinha(uint8_t pino) {
-	_pino = pino;
-
-	pinMode(_pino, INPUT);
-}
-
-int SensorDeLinha::ler() {
-	return analogRead(_pino);
-}
-
-// -------------------------------- Classe SensorDeToque ------------------------------------ //
-
-SensorDeToque::SensorDeToque(uint8_t porta) {
-	_porta = porta;
-	pinMode(_porta, INPUT_PULLUP);
-}
-
-boolean SensorDeToque::lerEstado() {
-	return !digitalRead(_porta);
-}
-
-// ----------------------------- Classe SensorDeDistancia ----------------------------------- //
-
-SensorDeDistancia::SensorDeDistancia(uint8_t porta) {
-	_porta = porta;
-
-	switch (_porta) {
-	case L1:
-		shut = SHUT_L1;
-		endereco = ENDERECO_L1;
-		break;
-	case L2:
-		shut = SHUT_L2;
-		endereco = ENDERECO_L2;
-		break;
-	case L3:
-		shut = SHUT_L3;
-		endereco = ENDERECO_L3;
-		break;
-	case L4:
-		shut = SHUT_L4;
-		endereco = ENDERECO_L4;
-		break;
-	case L5:
-		shut = SHUT_L5;
-		endereco = ENDERECO_L5;
-		break;
-	case L6:
-		shut = SHUT_L6;
-		endereco = ENDERECO_L6;
-		break;
-	case L7:
-		shut = SHUT_L7;
-		endereco = ENDERECO_L7;
-		break;
-	default:
-		break;
-	}
-
-	pinMode(shut, OUTPUT);
-	digitalWrite(shut, LOW);
-}
-
-boolean SensorDeDistancia::iniciar() {
-	digitalWrite(shut, HIGH);
-  delay(10);
-
-	sensor.begin(endereco);
-}
-
-int SensorDeDistancia::ler() {
-	sensor.rangingTest(&distancia);
-
-  if (distancia.RangeStatus != 4) {
-    return distancia.RangeMilliMeter/10;
-  }
-  return 819;
-}
-
 // ----------------------------------- Classe SensorDeCor ----------------------------------- //
 
 SensorDeCor::SensorDeCor(uint8_t porta) 
 {
 	switch (porta) {
 		case D:
-			_sdaPin = SDAD;
-			_sclPin = SCLD;
+			_sdaPin = PINO_SDA_SENSOR_DE_COR_DIREITO;
+			_sclPin = PINO_SCL_SENSOR_DE_COR_DIREITO;
 			break;
 		case E:
-			_sdaPin = SDAE;
-			_sclPin = SCLE;
+			_sdaPin = PINO_SDA_SENSOR_DE_COR_ESQUERDO;
+			_sclPin = PINO_SCL_SENSOR_DE_COR_ESQUERDO;
 			break;
 		default:
 			return;
@@ -342,177 +199,3 @@ void SensorDeCor::obterValores (uint16_t *r, uint16_t *g, uint16_t *b, uint16_t 
       break;
   }
 }
-
-// ------------------------------------ Classe Buzzer --------------------------------------- //
-
-Buzzer::Buzzer() {
-	pinMode(BUZZER, OUTPUT);
-}
-
-void Buzzer::soar_bipe() {
-	tone(BUZZER, 1000);
-  delay(200);
-  noTone(BUZZER);
-}
-
-void Buzzer::soar_dois_bipes() {
-	soar_bipe();
-	delay(100);
-	soar_bipe();
-}
-
-// ------------------------------------- Classe Motor --------------------------------------- //
-
-Motor::Motor(uint8_t porta) {
-  _porta = porta;
-
-  switch (_porta) {
-	    case M1:
-	      pinMode(MOTOR1_A, OUTPUT);
-	      pinMode(MOTOR1_B, OUTPUT);
-	      pinMode(PINO_VELOCIDADE_MOTOR_1, OUTPUT);
-	      break;
-	    case M2:
-	      pinMode(MOTOR2_A, OUTPUT);
-	      pinMode(MOTOR2_B, OUTPUT);
-	      pinMode(PINO_VELOCIDADE_MOTOR_2, OUTPUT);
-	      break;
-      default:
-        break;
-	}
-
-	configurarVelocidade(100);
-}
-
-void Motor::configurarVelocidade(uint8_t valor) {
-  _velocidade = valor;
-	
-	switch (_porta) {
-	    case M1:
-	      analogWrite(PINO_VELOCIDADE_MOTOR_1, _velocidade);
-	      break;
-	    case M2:
-	      analogWrite(PINO_VELOCIDADE_MOTOR_2, _velocidade);
-	      break;
-	}
-}
-
-void Motor::girar(uint8_t sentido) {
-	_sentido = sentido;
-	
-	switch (_porta) {
-	    case M1:
-	      a = MOTOR1_A; b = MOTOR1_B;
-	      break;
-	    case M2:
-	      a = MOTOR2_A; b = MOTOR2_B;
-	      break;
-	    default:
-	      return;
-	}
-
-	switch (_sentido) {
-	    case PARA_FRENTE:
-	      digitalWrite(a, HIGH);
-	      digitalWrite(b, LOW);
-	      break;
-	    case PARA_TRAS:
-	      digitalWrite(a, LOW);
-	      digitalWrite(b, HIGH);
-	      break;
-			default:
-				break;
-	}
-}
-
-void Motor::parar() {
-	switch (_porta) {
-	    case 1:
-	      a = MOTOR1_A; b = MOTOR1_B;
-	      break;
-	    case 2:
-	      a = MOTOR2_A; b = MOTOR2_B;
-	      break;
-	    default:
-	      return;
-	}
-
-	digitalWrite(a, LOW);
-	digitalWrite(b, LOW);
-}
-
-// ----------------------------------- Classe ServoMotor ------------------------------------ //
-
-ServoMotor::ServoMotor(uint8_t porta) {
-	servo.attach(porta);
-}
-
-void ServoMotor::anexar(uint8_t porta) {
-	servo.attach(porta);
-}
-
-void ServoMotor::desanexar() {
-	servo.detach();
-}
-
-void ServoMotor::gravar(uint8_t angulo) {
-	servo.write(angulo);
-}
-
-// ----------------------------------- Classe Tela --------------------------------------- //
-
-Tela::Tela() {
-  display = Adafruit_SSD1306(128, 64, &Wire, 4);
-}
-
-void Tela::iniciar() {
-  display.begin(SSD1306_SWITCHCAPVCC, 0x3C);
-}
-
-void Tela::configurarCursor(uint16_t x, uint16_t y) {
-  display.setCursor(x, y);
-}
-
-void Tela::mostrar() {
-  display.display();
-}
-
-void Tela::limpar() {
-  display.clearDisplay();
-}
-
-void Tela::configurarTamanhoDoTexto(uint8_t tamanho) {
-  display.setTextSize(tamanho);
-}
-
-void Tela::configurarCorDoTexto(uint16_t cor) {
-  display.setTextColor(cor);
-}
-
-void Tela::escrever(const char* texto) {
-  display.print(texto);
-}
-
-void Tela::escrever(int numero) {
-  display.print(numero);
-}
-
-void Tela::escrever(float numero) {
-  display.print(numero);
-}
-
-void Tela::escrever(long numero) {
-  display.print(numero);
-}
-
-void Tela::escrever(unsigned long numero) {
-  display.print(numero);
-}
-
-void Tela::desenharImagem(uint16_t x, uint16_t y, const uint8_t imagem[], uint16_t altura, 
-                             uint16_t largura, uint16_t cor) 
-{
-  display.drawBitmap(x, y, imagem, largura, altura, cor);
-}
-
-// ---------------------------------------- FIM DO CÓDIGO ----------------------------------- //
